use new 'unique realloc' of ID.
that way we do not have to stupidly ensure unique addresses of all IDs
read during an undo step, but only for those we actually need.

Combined with the 'almost always reuse old address' feature of undo swap
branch, this should ensure us to almost never call that code in
practice.