IO: ensure export path and export name are always consistent
Before this, there was one code path that set `context.export_path`,
and a different code path for `context.export_name`, allowing the two to
diverge.

Keeping track of the export path of the export parent (which can be but is
not always the Blender parent object) also allows a concrete subclass of
`AbstractHierarchyIterator` to find the `AbstractWriter` for the export
parent. In case of exporting to Alembic this is important, as it's not
possible to simply give the Alembic library the full export path of an
object (it needs the C++ object of the parent).